Save code of registered custom objects
